+++ packages/vite-plugin-angular/src/lib/nx-folder-plugin.ts
@@ -1,0 +1,19 @@
+import { normalizePath, Plugin } from 'vite';
+
+/**
+ * Ignores anything in the .nx folder from triggering HMR
+ *
+ * @returns
+ */
+export function nxFolderPlugin(): Plugin {
+  return {
+    name: 'analogjs-nx-folder-plugin',
+    handleHotUpdate(ctx) {
+      if (ctx.file.includes(normalizePath('/.nx/'))) {
+        return [];
+      }
+
+      return ctx.modules;
+    },
+  };
+}